test: Organize regression tests into categorized folder structure
