+++ packages/testing/playwright/tests/ui/regression/canvas/ADO-2929-can-load-old-switch-node-workflows.spec.ts
@@ -1,5 +1,5 @@
-import { test, expect } from '../../fixtures/base';
-import type { TestRequirements } from '../../Types';
+import { test, expect } from '../../../../fixtures/base';
+import type { TestRequirements } from '../../../../Types';
 
 const requirements: TestRequirements = {
 	workflow: {
