+++ packages/testing/playwright/tests/ui/regression/ndv/ADO-1338-ndv-missing-input-panel.spec.ts
@@ -1,5 +1,5 @@
-import { test, expect } from '../../fixtures/base';
-import type { TestRequirements } from '../../Types';
+import { test, expect } from '../../../../fixtures/base';
+import type { TestRequirements } from '../../../../Types';
 
 const requirements: TestRequirements = {
 	workflow: {
