+++ packages/testing/playwright/tests/ui/regression/ndv/ADO-2270-opening-webhook-ndv-marks-workflow-as-unsaved.spec.ts
@@ -1,4 +1,4 @@
-import { test, expect } from '../../fixtures/base';
+import { test, expect } from '../../../../fixtures/base';
 
 test.describe('ADO-2270 Save button resets on webhook node open', () => {
 	test('should not reset the save button if webhook node is opened and closed', async ({ n8n }) => {
