+++ packages/testing/playwright/tests/ui/regression/ndv/ADO-2372-prevent-clipping-params.spec.ts
@@ -1,4 +1,4 @@
-import { test, expect } from '../../fixtures/base';
+import { test, expect } from '../../../../fixtures/base';
 
 test.describe('ADO-2362 ADO-2350 NDV Prevent clipping long parameters and scrolling to expression', () => {
 	test('should show last parameters and open at scroll top of parameters', async ({ n8n }) => {
