+++ packages/testing/playwright/tests/ui/regression/ndv/SUG-121-fields-reset-after-closing-ndv.spec.ts
@@ -1,5 +1,5 @@
-import { test, expect } from '../../fixtures/base';
-import type { TestRequirements } from '../../Types';
+import { test, expect } from '../../../../fixtures/base';
+import type { TestRequirements } from '../../../../Types';
 
 const requirements: TestRequirements = {
 	workflow: 'Test_workflow_1.json',
